+++ tasks/src/robocorp/tasks/_constants.py
@@ -1,0 +1,2 @@
+# TODO: Support array and object.
+SUPPORTED_TYPES_IN_SCHEMA = (str, int, float, bool)